requirements.txt
pyyaml~=6.0
numpy~=1.25.0
pandas~=2.0.2
openai~=1.3.2
python-dotenv~=1.0.0
lxml~=4.9.3
rdflib~=6.3.2
morph-kgc==2.6.4
scikit-learn~=1.2.2
sentence_transformers~=2.7.0
PyQt5~=5.15.9
markdown~=3.5
matplotlib~=3.7.1
requests~=2.31.0
Pillow~=9.5.0
networkx~=2.8.8
pyvis~=0.3.2
setuptools
nltk~=3.8.1
wordcloud~=1.9.2
beautifulsoup4~=4.12.2
python-dotenv
serpapi
chardet